+++ testcases/functional_testcases/test_conversion_handlers.py
@@ -1,14 +1,12 @@
 import json
 import unittest
 from decimal import Decimal
-from unittest import mock
 from unittest.mock import patch, Mock
 
 from sqlalchemy import distinct
 
-import utils.blockchain
 from application.handler.conversion_handlers import create_conversion_request, get_conversion_history, \
-    create_transaction_for_conversion, claim_conversion, get_all_deposit_address
+    create_transaction_for_conversion, claim_conversion, get_conversion
 from constants.error_details import ErrorCode, ErrorDetails
 from constants.lambdas import LambdaResponseStatus
 from constants.status import ConversionStatus
@@ -46,7 +44,8 @@ class TestConversion(unittest.TestCase):
         conversion_repo.session.add_all(TestVariables().transaction)
         conversion_repo.session.commit()
 
-    @patch("utils.blockchain.get_cardano_deposit_address", Mock(return_value="some derived address"))
+    @patch("application.service.cardano_service.CardanoService.get_deposit_address",
+           Mock(return_value={"derived_address": "some derived address", "index": 1, "role": 0}))
     @patch("application.service.conversion_service.get_signature")
     @patch("utils.blockchain.validate_cardano_address")
     @patch("common.utils.Utils.report_slack")
@@ -261,29 +260,24 @@ class TestConversion(unittest.TestCase):
             'id': '7298bce110974411b260cac758b37ee0', 'deposit_amount': '1.33305E+8', 'claim_amount': None,
             'fee_amount': None, 'status': 'USER_INITIATED', 'updated_at': '2022-01-12 04:10:54'}, 'wallet_pair': {
             'from_address': '0xa18b95A9371Ac18C233fB024cdAC5ef6300efDa1',
-            'to_address': 'addr_test1qza8485avt2xn3vy63plawqt0gk3ykpf98wusc4qrml2avu0pkm5rp3pkz6q4n3kf8znlf3y749lll8lfmg5x86kgt8qju7vx8'},
-            'from_token': {
-                'name': 'Singularity Ethereum',
-                'symbol': 'AGIX',
-                'blockchain': {'name': 'Ethereum',
-                               'symbol': 'ETH',
-                               'chain_id': 42}},
-            'to_token': {
-                'name': 'Singularity Cardano',
-                'symbol': 'AGIX',
-                'blockchain': {'name': 'Cardano',
-                               'symbol': 'ADA',
-                               'chain_id': 2}},
-            'transactions': []}, {'conversion': {
+            'to_address': 'addr_test1qza8485avt2xn3vy63plawqt0gk3ykpf98wusc4qrml2avu0pkm5rp3pkz6q4n3kf8znlf3y749lll8lfmg5x86kgt8qju7vx8',
+            'deposit_address': None}, 'from_token': {'name': 'Singularity Ethereum', 'symbol': 'AGIX',
+                                                     'blockchain': {'name': 'Ethereum', 'symbol': 'ETH',
+                                                                    'chain_id': 42}}, 'to_token': {
+            'name': 'Singularity Cardano', 'symbol': 'AGIX',
+            'blockchain': {'name': 'Cardano', 'symbol': 'ADA', 'chain_id': 2}}, 'transactions': []}, {'conversion': {
             'id': '5086b5245cd046a68363d9ca8ed0027e', 'deposit_amount': '1.33305E+18', 'claim_amount': None,
             'fee_amount': None, 'status': 'USER_INITIATED', 'updated_at': '2022-01-12 04:10:54'}, 'wallet_pair': {
             'from_address': 'addr_test1qza8485avt2xn3vy63plawqt0gk3ykpf98wusc4qrml2avu0pkm5rp3pkz6q4n3kf8znlf3y749lll8lfmg5x86kgt8qju7vx8',
-            'to_address': '0xa18b95A9371Ac18C233fB024cdAC5ef6300efDa1'}, 'from_token': {'name': 'Singularity Cardano',
-                                                                                        'symbol': 'AGIX',
-                                                                                        'blockchain': {
-                                                                                            'name': 'Cardano',
-                                                                                            'symbol': 'ADA',
-                                                                                            'chain_id': 2}},
+            'to_address': '0xa18b95A9371Ac18C233fB024cdAC5ef6300efDa1',
+            'deposit_address': 'addr_test1qza8485avt2xn3vy63plawqt0gk3ykpf98wusc4qrml2avu0pkm5rp3pkz6q4n3kf8znlf3y749lll8lfmg5x86kgt8qju7vx8'},
+            'from_token': {
+                'name': 'Singularity Cardano',
+                'symbol': 'AGIX',
+                'blockchain': {
+                    'name': 'Cardano',
+                    'symbol': 'ADA',
+                    'chain_id': 2}},
             'to_token': {
                 'name': 'Singularity Ethereum',
                 'symbol': 'AGIX',
@@ -301,7 +295,8 @@ class TestConversion(unittest.TestCase):
                 'updated_at': '2022-01-12 04:10:54'},
                 'wallet_pair': {
                     'from_address': '0xa18b95A9371Ac18C233fB024cdAC5ef6300efDa1',
-                    'to_address': 'addr_test1qza8485avt2xn3vy63plawqt0gk3ykpf98wusc4qrml2avu0pkm5rp3pkz6q4n3kf8znlf3y749lll8lfmg5x86kgt8qju7vx8'},
+                    'to_address': 'addr_test1qza8485avt2xn3vy63plawqt0gk3ykpf98wusc4qrml2avu0pkm5rp3pkz6q4n3kf8znlf3y749lll8lfmg5x86kgt8qju7vx8',
+                    'deposit_address': None},
                 'from_token': {
                     'name': 'Singularity Ethereum',
                     'symbol': 'AGIX',
@@ -362,6 +357,51 @@ class TestConversion(unittest.TestCase):
         response = get_conversion_history(event, {})
         body = json.loads(response["body"])
         self.assertEqual(body, success_response_with_no_history)
+
+    @patch("common.utils.Utils.report_slack")
+    def test_get_conversion(self, mock_report_slack):
+        event = dict()
+
+        bad_request_property_value_empty = {'status': 'failed', 'data': None,
+                                            'error': {'code': 'E0005', 'message': 'BAD_REQUEST',
+                                                      'details': 'Property value is empty'}}
+        bad_request_invalid_conversion_id = {'status': 'failed', 'data': None,
+                                             'error': {'code': 'E0008', 'message': 'BAD_REQUEST',
+                                                       'details': 'Invalid conversion id provided'}}
+        success_response = {'status': 'success', 'data': {
+            'conversion': {'id': '51769f201e46446fb61a9c197cb0706b', 'deposit_amount': '1.66305E+18',
+                           'claim_amount': None, 'fee_amount': None, 'status': 'PROCESSING',
+                           'updated_at': '2022-01-12 04:10:54'},
+            'wallet_pair': {'from_address': '0xa18b95A9371Ac18C233fB024cdAC5ef6300efDa1',
+                            'to_address': 'addr_test1qza8485avt2xn3vy63plawqt0gk3ykpf98wusc4qrml2avu0pkm5rp3pkz6q4n3kf8znlf3y749lll8lfmg5x86kgt8qju7vx8',
+                            'deposit_address': None}, 'from_token': {'name': 'Singularity Ethereum', 'symbol': 'AGIX',
+                                                                     'blockchain': {'name': 'Ethereum', 'symbol': 'ETH',
+                                                                                    'chain_id': 42}},
+            'to_token': {'name': 'Singularity Cardano', 'symbol': 'AGIX',
+                         'blockchain': {'name': 'Cardano', 'symbol': 'ADA', 'chain_id': 2}}, 'transactions': [
+                {'id': '391be6385abf4b608bdd20a44acd6abc', 'transaction_operation': 'TOKEN_RECEIVED',
+                 'transaction_hash': '22477fd4ea994689a04646cbbaafd133', 'transaction_amount': '1.66305E+18',
+                 'status': 'SUCCESS', 'updated_at': '2022-01-12 04:10:54'},
+                {'id': '1df60a2369f34247a5dc3ed29a8eef67', 'transaction_operation': 'TOKEN_RECEIVED',
+                 'transaction_hash': '22477fd4ea994689a04646cbbaafd133', 'transaction_amount': '1.66305E+18',
+                 'status': 'WAITING_FOR_CONFIRMATION', 'updated_at': '2022-01-12 04:10:54'}]},
+                            'error': {'code': None, 'message': None, 'details': None}}
+
+        response = get_conversion(event, {})
+        body = json.loads(response["body"])
+        self.assertEqual(body, bad_request_property_value_empty)
+
+        response = get_conversion({"pathParameters": {"conversion_id": ""}}, {})
+        body = json.loads(response["body"])
+        self.assertEqual(body, bad_request_property_value_empty)
+
+        response = get_conversion({"pathParameters": {"conversion_id": "random conversion id "}}, {})
+        body = json.loads(response["body"])
+        self.assertEqual(body, bad_request_invalid_conversion_id)
+
+        response = get_conversion({"pathParameters": {"conversion_id": "51769f201e46446fb61a9c197cb0706b"}}, {})
+        body = json.loads(response["body"])
+        self.assertEqual(body, success_response)
 
     @patch("utils.blockchain.get_cardano_transaction_details")
     @patch("utils.blockchain.get_ethereum_transaction_details")
@@ -600,25 +640,6 @@ class TestConversion(unittest.TestCase):
         response = claim_conversion(event, {})
         body = json.loads(response["body"])
         self.assertEqual(body, success_response)
-
-    @patch("common.utils.Utils.report_slack")
-    def test_get_all_deposit_address(self, mock_report_slack):
-        event = dict()
-        success_response = {'status': 'success', 'data': {'addresses': [
-            'addr_test1qza8485avt2xn3vy63plawqt0gk3ykpf98wusc4qrml2avu0pkm5rp3pkz6q4n3kf8znlf3y749lll8lfmg5x86kgt8qju7vx8']},
-                            'error': {'code': None, 'message': None, 'details': None}}
-        success_response_no_addresses = {'status': 'success', 'data': {'addresses': []},
-                                         'error': {'code': None, 'message': None, 'details': None}}
-
-        response = get_all_deposit_address(event, {})
-        body = json.loads(response["body"])
-        self.assertEqual(body, success_response)
-
-        TestConversion.delete_all_tables()
-
-        response = get_all_deposit_address(event, {})
-        body = json.loads(response["body"])
-        self.assertEqual(body, success_response_no_addresses)
 
     def tearDown(self):
         TestConversion.delete_all_tables()